Move drawer up to DefaultLayout (page level). Need a solution to control it from the table components. context/portals?
