+++ client/src/app/shared/components/app-table-with-controls/app-table-with-controls.tsx
@@ -44,76 +44,50 @@ export const AppTableWithControls: React.FC<IAppTableWithControlsProps> = ({
   toolbarClearAllFilters,
   paginationProps,
   paginationIdPrefix,
-  activeAppInDetailDrawer,
-  closeDetailDrawer,
+  activeAppInDetailDrawer, // TODO this is unused for now since we lifted the drawer to DefaultLayout. How to wire it up? context? portals?
+  closeDetailDrawer, // TODO this is unused for now since we lifted the drawer to DefaultLayout. How to wire it up? context? portals?
   ...rest
 }) => {
   const { t } = useTranslation();
 
-  const drawerRef = React.useRef<HTMLSpanElement>(null);
-
-  const drawerPanelContent = (
-    <DrawerPanelContent
-      isResizable
-      id="app-detail-drawer"
-      defaultSize="500px"
-      minSize="150px"
-    >
-      <DrawerHead>
-        <span tabIndex={activeAppInDetailDrawer ? 0 : -1} ref={drawerRef}>
-          TODO - drawer header here!
-        </span>
-      </DrawerHead>
-      <DrawerActions>
-        <DrawerCloseButton onClick={closeDetailDrawer} />
-      </DrawerActions>
-    </DrawerPanelContent>
-  );
-
   return (
-    <Drawer isExpanded={!!activeAppInDetailDrawer} position="right">
-      <DrawerContent panelContent={drawerPanelContent}>
-        <DrawerContentBody>
-          <div
-            style={{
-              backgroundColor: "var(--pf-global--BackgroundColor--100)",
-            }}
-          >
-            <Toolbar
-              className="pf-m-toggle-group-container"
-              collapseListedFiltersBreakpoint="xl"
-              clearAllFilters={toolbarClearAllFilters}
-              clearFiltersButtonText={t("actions.clearAllFilters")}
+    <div
+      style={{
+        backgroundColor: "var(--pf-global--BackgroundColor--100)",
+      }}
+    >
+      <Toolbar
+        className="pf-m-toggle-group-container"
+        collapseListedFiltersBreakpoint="xl"
+        clearAllFilters={toolbarClearAllFilters}
+        clearFiltersButtonText={t("actions.clearAllFilters")}
+      >
+        <ToolbarContent>
+          {toolbarBulkSelector}
+          {toolbarToggle ? toolbarToggle : null}
+          {toolbarActions}
+          {!withoutTopPagination && (
+            <ToolbarItem
+              variant={ToolbarItemVariant.pagination}
+              alignment={{ default: "alignRight" }}
             >
-              <ToolbarContent>
-                {toolbarBulkSelector}
-                {toolbarToggle ? toolbarToggle : null}
-                {toolbarActions}
-                {!withoutTopPagination && (
-                  <ToolbarItem
-                    variant={ToolbarItemVariant.pagination}
-                    alignment={{ default: "alignRight" }}
-                  >
-                    <SimplePagination
-                      idPrefix={paginationIdPrefix}
-                      isTop={true}
-                      paginationProps={paginationProps}
-                    />
-                  </ToolbarItem>
-                )}
-              </ToolbarContent>
-            </Toolbar>
-            <AppTable {...rest} />
-            {!withoutBottomPagination && (
               <SimplePagination
                 idPrefix={paginationIdPrefix}
-                isTop={false}
+                isTop={true}
                 paginationProps={paginationProps}
               />
-            )}
-          </div>
-        </DrawerContentBody>
-      </DrawerContent>
-    </Drawer>
+            </ToolbarItem>
+          )}
+        </ToolbarContent>
+      </Toolbar>
+      <AppTable {...rest} />
+      {!withoutBottomPagination && (
+        <SimplePagination
+          idPrefix={paginationIdPrefix}
+          isTop={false}
+          paginationProps={paginationProps}
+        />
+      )}
+    </div>
   );
 };
